#152: Refactoring dividing line styling options
