+++ src/app/parcels/parcelsTable/conditionalStyling.ts
@@ -1,13 +1,24 @@
+import { useTheme } from "styled-components";
 import { ParcelsTableRow } from "./types";
 import { BreakPointConfig } from "@/components/Tables/Table";
+
+export type DividingLineStyleOptions = {
+    dateAndSlotPrimary: DividingLineStyle;
+    dateAndSlotSecondary: DividingLineStyle;
+    slotPrimary: DividingLineStyle;
+};
+
+type DividingLineStyle = {
+    colour: string;
+    thickness: string;
+};
 
 const getRowBoundaries = (columnValues: (string | number | null)[]): number[] => {
     const boundaries: number[] = [];
 
     columnValues.forEach((columnValue, index) => {
         const previousIndex = index - 1 >= 0 ? index - 1 : 0;
-        const previousValue = columnValues[previousIndex];
-        if (columnValue !== previousValue) {
+        if (columnValue !== columnValues[previousIndex]) {
             boundaries.push(index);
         }
     });
@@ -56,3 +67,22 @@ export const searchForBreakPoints = (
             return [];
     }
 };
+
+export const getDividingLineStyleOptions = (): DividingLineStyleOptions => {
+    const theme = useTheme()
+    return (
+        {
+        dateAndSlotPrimary: {
+            colour: theme.primary.background[3],
+            thickness: "5pt",
+        },
+        dateAndSlotSecondary: {
+            colour: theme.primary.background[2],
+            thickness: "2.5pt",
+        },
+        slotPrimary: {
+            colour: theme.primary.background[3],
+            thickness: "5pt",
+        },
+    }
+)}
